Refactor: move repositories and entities to core module
