+++ app-module/src/main/java/com/alberto/app/person/service/impl/TeacherServiceImpl.java
@@ -2,8 +2,8 @@ package com.alberto.app.person.service.impl;
 
 import com.alberto.app.person.model.dto.TeacherDto;
 import com.alberto.app.person.model.mapper.TeacherMapper;
-import com.alberto.app.person.repository.TeacherRepository;
 import com.alberto.app.person.service.TeacherService;
+import com.alberto.core.person.repository.TeacherRepository;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
