+++ core-module/src/main/java/com/alberto/core/account/repository/AccountRepository.java
@@ -1,6 +1,6 @@
-package com.alberto.app.account.repository;
+package com.alberto.core.account.repository;
 
-import com.alberto.app.account.model.entity.AccountEntity;
+import com.alberto.core.account.model.entity.AccountEntity;
 import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 
 import javax.enterprise.context.ApplicationScoped;
